+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/rtc/PeerNetworkConfig.kt
@@ -1,12 +1,13 @@
 package jp.shiguredo.sora.sdk.channel.rtc
 
+import jp.shiguredo.sora.sdk.channel.option.SoraMediaOption
 import jp.shiguredo.sora.sdk.channel.signaling.message.OfferConfig
 import org.webrtc.MediaConstraints
 import org.webrtc.PeerConnection
 
 class PeerNetworkConfig(
         private val serverConfig: OfferConfig,
-        private val enableTcp:    Boolean = false
+        private val mediaOption:  SoraMediaOption
 ) {
     fun createConfiguration(): PeerConnection.RTCConfiguration {
 
@@ -23,12 +24,12 @@ class PeerNetworkConfig(
         conf.continualGatheringPolicy = PeerConnection.ContinualGatheringPolicy.GATHER_CONTINUALLY
         conf.keyType                  = PeerConnection.KeyType.ECDSA
 
-        conf.tcpCandidatePolicy = if (enableTcp)
-            PeerConnection.TcpCandidatePolicy.ENABLED
-        else
-            PeerConnection.TcpCandidatePolicy.DISABLED
-
         conf.enableDtlsSrtp = true;
+        conf.sdpSemantics = mediaOption.sdpSemantics
+
+        conf.tcpCandidatePolicy = mediaOption.tcpCandidatePolicy
+        conf.enableCpuOveruseDetection = mediaOption.enableCpuOveruseDetection
+
         return conf
     }